@@ -1,18 +1,16 @@
 {
-  "name": "react-folder-structure",
-  "homepage": "https://itskd.io",
-  "version": "1.1.0",
+  "name": "my-portfolio",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^13.5.0",
-    "emailjs-com": "^3.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-scripts": "5.0.0",
-    "swiper": "^8.0.6",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-scripts": "5.0.1",
+    "source-map-resolve": "^0.6.0",
+    "svgo": "^2.8.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
