@@ -1,29 +1,18 @@
-import React from 'react';
-import './index.css';
 import About from './components/about/About';
-import Header from './components/header/Header';
-import Nav from './components/nav/Nav';
-import Experience from './components/experience/Experience';
-import Services from './components/services/Services';
-import Portfolio from './components/portfolio/Portfolio';
-import Testimonials from './components/testimonials/Testimonials';
 import Contact from './components/contact/Contact';
-import Footer from './components/footer/Footer';
+import Intro from './components/intro/Intro'
+import ProjectList from './components/projectList/ProjectList';
 
-const App = () => {
+
+function App() {
   return (
-    <>
-        <Header />
-        <Nav />
-        <About />
-        <Experience />
-        <Services />
-        <Portfolio />
-        <Testimonials />
-        <Contact />
-        <Footer />
-    </>
-  )
+    <div className="App">
+      <Intro />
+      <About />
+      <ProjectList />
+      <Contact />
+    </div>
+  );
 }
 
-export default App
+export default App;
